Create word list from backend
